requirements.dev.txt
pytest==7.3.1
pytest-aiohttp==1.0.4
